Fix AcoustId Response
Suddenly the response moved to float, before it was always whole numbers

+++ MusicMover/Models/AcoustIdRecordingResponse.cs
@@ -3,7 +3,7 @@ namespace MusicMover.Models;
 public class AcoustIdRecordingResponse
 {
     public string? Id { get; set; }
-    public int? Duration { get; set; }
+    public float? Duration { get; set; }
     public string? Title { get; set; }
     public List<AcoustIdArtistsResponse>? Artists { get; set; }
 }
